api/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
ultralytics==8.0.196
opencv-python==4.8.1.78
Pillow==10.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
PyYAML==6.0.1
